+++ packages/configs/other/prettier.config.js
@@ -1,0 +1,10 @@
+module.exports = {
+	useTabs: true,
+	singleQuote: true,
+	semi: true,
+	proseWrap: 'always',
+	trailingComma: 'all',
+	printWidth: 80,
+	arrowParens: 'always',
+	bracketSameLine: false,
+};