+++ packages/create-pantheon-decoupled-kit/src/generators/next-wp.generator.ts
@@ -1,4 +1,5 @@
 import { addWithDiff, runInstall, runLint } from '../actions';
+import versions from '../pkgVersions.json';
 import type { DecoupledKitGenerator, DefaultAnswers } from '../types';
 
 interface NextWPAnswers extends DefaultAnswers {
@@ -21,6 +22,10 @@ export const nextWp: DecoupledKitGenerator<NextWPAnswers> = {
 				`${process.cwd()}/${appName.replaceAll(' ', '-').toLowerCase()}`,
 		},
 	],
+	data: {
+		nextjsKitVersion: versions['nextjs-kit'],
+		wordpressKitVersion: versions['wordpress-kit'],
+	},
 	templates: ['next-wp'],
 	actions: [addWithDiff, runInstall, runLint],
 };