+++ packages/create-pantheon-decoupled-kit/CHANGELOG.md
@@ -1,5 +1,37 @@
 # create-pantheon-decoupled-kit
 
+## 0.4.0-canary.3
+
+### Patch Changes
+
+- 2cc371a: Updates to watch script
+- b719c1c: Add dynamic versioning for @pantheon-systems dependencies
+- d38e923: Removed .npmignore from project templates
+
+## 0.4.0-canary.2
+
+### Patch Changes
+
+- 1218c02: [gatsby-wp] Add conditinal to gatsby-config to avoid loading missing
+  plugins
+
+## 0.4.0-canary.1
+
+### Minor Changes
+
+- 981825d: [gatsby-wp] Add Published `eslint` Configs to Gatsby Generator
+
+### Patch Changes
+
+- c31795b: Rename gitignore template on write to workaround npm
+
+## 0.3.1-canary.0
+
+### Patch Changes
+
+- 9ede7c8: Update config package route.
+- 4dd905b: [next-wp] Update post and page date
+
 ## 0.3.0
 
 ### Minor Changes